refactor: check device hook
