+++ src/encrypted_notes_frontend/src/hooks/useDeviceCheck.ts
@@ -1,45 +1,20 @@
-import { useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
+import { useCallback } from 'react';
 
 import { useAuthContext } from './authContext';
-import { useMessage } from './useMessage';
 
 export const useDeviceCheck = () => {
-  const navigate = useNavigate();
-  const { auth, logout } = useAuthContext();
-  const { showMessage } = useMessage();
+  const { auth } = useAuthContext();
 
-  useEffect(() => {
-    // 1秒ごとにポーリングします。
-    const intervalId = setInterval(async () => {
-      console.log('Check device data...');
-      if (auth.status !== 'SYNCED') {
-        return;
-      }
+  const isDeviceRemoved = useCallback(async () => {
+    if (auth.status !== 'SYNCED') {
+      return false;
+    }
 
-      const deviceAlias = await auth.actor.getDeviceAliases();
-      if (
-        !deviceAlias.find(
-          (deviceAlias) => deviceAlias === auth.cryptoService.deviceAlias,
-        )
-      ) {
-        try {
-          await logout();
-          showMessage({
-            title: 'This device has been deleted.',
-            status: 'info',
-          });
-          navigate('/');
-        } catch (err) {
-          showMessage({ title: 'Failed to logout', status: 'error' });
-          console.error(err);
-        }
-      }
-    }, 1000);
-
-    // クリーンアップ関数を返します。
-    return () => {
-      clearInterval(intervalId);
-    };
+    // バックエンドキャニスターからデバイスエイリアスを取得します。
+    const deviceAlias = ['dummy'];
+    // 自身のデバイスエイリアスが含まれていない場合は、デバイスが削除されたと判断します。
+    return !deviceAlias.includes(auth.cryptoService.deviceAlias);
   }, [auth]);
+
+  return { isDeviceRemoved };
 };
