+++ src/encrypted_notes_frontend/src/routes/devices/index.tsx
@@ -7,17 +7,25 @@ import { useDeviceCheck, useMessage } from '../../hooks';
 import { useAuthContext } from '../../hooks/authContext';
 
 export const Devices = () => {
-  const { auth } = useAuthContext();
-  const { showMessage } = useMessage();
   const {
     isOpen: isOpenDeleteDialog,
     onOpen: onOpenDeleteDialog,
     onClose: onCloseDeleteDialog,
   } = useDisclosure();
   const navigate = useNavigate();
+
+  const { auth, logout } = useAuthContext();
+  const { isDeviceRemoved } = useDeviceCheck();
+  const { showMessage } = useMessage();
+
   const [deviceAliases, setDeviceAliases] = useState<string[]>([]);
   const [deleteAlias, setDeleteAlias] = useState<string | undefined>(undefined);
   const [isLoading, setIsLoading] = useState(false);
+
+  const openDeleteDialog = (alias: string) => {
+    setDeleteAlias(alias);
+    onOpenDeleteDialog();
+  };
 
   const deleteDevice = async () => {
     if (auth.status !== 'SYNCED') {
@@ -68,6 +76,47 @@ export const Devices = () => {
       await getDevices();
     })();
   }, [auth]);
+
+  useEffect(() => {
+    // 1秒ごとにポーリングします。
+    const intervalId = setInterval(async () => {
+      console.log('Check device data...');
+
+      const isRemoved = await isDeviceRemoved();
+      if (isRemoved) {
+        try {
+          await logout();
+          showMessage({
+            title: 'This device has been deleted.',
+            status: 'info',
+          });
+          navigate('/');
+        } catch (err) {
+          showMessage({ title: 'Failed to logout', status: 'error' });
+          console.error(err);
+        }
+      }
+    }, 1000);
+
+    // クリーンアップ関数を返します。
+    return () => {
+      clearInterval(intervalId);
+    };
+  }, [auth]);
+
+  if (auth.status === 'ANONYMOUS') {
+    return null;
+  }
+
+  if (auth.status === 'SYNCHRONIZING') {
+    return (
+      <Layout>
+        <Box p={6} overflowY={'auto'} maxHeight={'calc(100vh - 64px)'}>
+          Loading...
+        </Box>
+      </Layout>
+    );
+  }
 
   return (
     <>
